Refactoring from PR comments
