+++ lib/responseDownloadFormats/html-aggregated/components/ResponseHTMLAggregated.tsx
@@ -1,8 +1,5 @@
 import React from "react";
-import {
-  CopyToClipboardScript,
-  copyCodeAndResponseFromTableToClipboardScript,
-} from "../../html/scripts";
+import { CopyToClipboardScript } from "../../html/scripts";
 import { ProtectedWarning } from "../../html/components/ProtectedWarning";
 import Fip from "../../html/components/Fip";
 import { css } from "../../html/css/compiled";
@@ -10,6 +7,7 @@ import { ColumnTable } from "../../html/components/ColumnTable";
 import { AggregatedTable } from "./AggregatedTable";
 import { FormResponseSubmissions, Submission } from "@lib/responseDownloadFormats/types";
 import { customTranslate } from "@lib/responseDownloadFormats/helpers";
+import { copyCodeAndResponseFromTableToClipboardScript } from "../scripts";
 
 interface HTMLDownloadProps {
   lang: string;
